+++ incubator/catalog-filesystem/src/main/java/io/aklivity/zilla/runtime/catalog/filesystem/internal/FilesystemEventContext.java
@@ -20,9 +20,9 @@ import java.time.Clock;
 import org.agrona.MutableDirectBuffer;
 import org.agrona.concurrent.UnsafeBuffer;
 
-import io.aklivity.zilla.incubator.catalog.filesystem.internal.types.event.EventFW;
-import io.aklivity.zilla.incubator.catalog.filesystem.internal.types.event.FilesystemEventExFW;
-import io.aklivity.zilla.incubator.catalog.filesystem.internal.types.event.FilesystemEventType;
+import io.aklivity.zilla.runtime.catalog.filesystem.internal.types.event.EventFW;
+import io.aklivity.zilla.runtime.catalog.filesystem.internal.types.event.FilesystemEventExFW;
+import io.aklivity.zilla.runtime.catalog.filesystem.internal.types.event.FilesystemEventType;
 import io.aklivity.zilla.runtime.engine.EngineContext;
 import io.aklivity.zilla.runtime.engine.binding.function.MessageConsumer;
